requirements.txt
django==5.1.3
Pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.1
djangorestframework==3.15.2